+++ src/planner/vectorfield/detail/VectorFieldIntegrator.cpp
@@ -11,8 +11,6 @@ namespace planner {
 namespace vectorfield {
 namespace detail {
 
-constexpr double defaultConstraintCheckResolution = 1e-3;
-
 //==============================================================================
 std::unique_ptr<aikido::trajectory::Spline> convertToSpline(
     const std::vector<Knot>& knots,
@@ -155,7 +153,7 @@ void VectorFieldIntegrator::check(const Eigen::VectorXd& q, double t)
   }
 }
 
-} // namesapce detail
+} // namespace detail
 } // namespace vectorfield
 } // namespace planner
 } // namespace aikido