+++ tests/constraint/test_CollisionFree.cpp
@@ -57,13 +57,13 @@ protected:
     mCollisionGroup3 = mCollisionDetector->createCollisionGroup(bn1, boxNode);
 
     // StateSpace for controlling skeleton
-    GroupPtr group = Group::create();
-    group->addBodyNode(bn1);
-    group->addBodyNode(boxNode);
-    group->addDofs(mManipulator->getDofs());
-    group->addDofs(mBox->getDofs());
+    mSkeleton = Group::create();
+    mSkeleton->addBodyNode(bn1);
+    mSkeleton->addBodyNode(boxNode);
+    mSkeleton->addDofs(mManipulator->getDofs());
+    mSkeleton->addDofs(mBox->getDofs());
 
-    mStateSpace = std::make_shared<MetaSkeletonStateSpace>(group);
+    mStateSpace = std::make_shared<MetaSkeletonStateSpace>(mSkeleton.get());
     mManipulator->enableSelfCollisionCheck();
     mManipulator->enableAdjacentBodyCheck();
   }
@@ -80,31 +80,34 @@ public:
 
   // statespace setup
   MetaSkeletonStateSpacePtr mStateSpace;
+  dart::dynamics::GroupPtr mSkeleton;
 };
 
 TEST_F(CollisionFreeTest, ConstructorThrowsOnNullStateSpace)
 {
   EXPECT_THROW(
-      CollisionFree(nullptr, mCollisionDetector), std::invalid_argument);
+      CollisionFree(nullptr, nullptr, mCollisionDetector),
+      std::invalid_argument);
 }
 
 TEST_F(CollisionFreeTest, ConstructorThrowsOnNullCollisionDetector)
 {
-  EXPECT_THROW(CollisionFree(mStateSpace, nullptr), std::invalid_argument);
+  EXPECT_THROW(
+      CollisionFree(mStateSpace, mSkeleton, nullptr), std::invalid_argument);
 }
 
 TEST_F(CollisionFreeTest, GetStateSpaceMatchStateSpace)
 {
-  CollisionFree constraint(mStateSpace, mCollisionDetector);
+  CollisionFree constraint(mStateSpace, mSkeleton, mCollisionDetector);
   EXPECT_EQ(mStateSpace, constraint.getStateSpace());
 }
 
 TEST_F(CollisionFreeTest, EmptyCollisionGroup_IsSatisfiedReturnsTrue)
 {
-  CollisionFree constraint(mStateSpace, mCollisionDetector);
+  CollisionFree constraint(mStateSpace, mSkeleton, mCollisionDetector);
 
-  auto state = mStateSpace->getScopedStateFromMetaSkeleton();
-  mStateSpace->setState(state);
+  auto state = mStateSpace->getScopedStateFromMetaSkeleton(mSkeleton.get());
+  mStateSpace->setState(mSkeleton.get(), state);
 
   std::unique_ptr<TestableOutcome> outcome = constraint.createOutcome();
   EXPECT_TRUE(constraint.isSatisfied(state, outcome.get()));
@@ -113,15 +116,15 @@ TEST_F(CollisionFreeTest, EmptyCollisionGroup_IsSatisfiedReturnsTrue)
 
 TEST_F(CollisionFreeTest, AddPairwiseCheckPasses_IsSatisfied)
 {
-  CollisionFree constraint(mStateSpace, mCollisionDetector);
+  CollisionFree constraint(mStateSpace, mSkeleton, mCollisionDetector);
   constraint.addPairwiseCheck(mCollisionGroup1, mCollisionGroup2);
 
-  auto state = mStateSpace->getScopedStateFromMetaSkeleton();
+  auto state = mStateSpace->getScopedStateFromMetaSkeleton(mSkeleton.get());
 
   Eigen::VectorXd position(Eigen::VectorXd::Zero(7));
   position(4) = 5;
   mStateSpace->convertPositionsToState(position, state);
-  mStateSpace->setState(state);
+  mStateSpace->setState(mSkeleton.get(), state);
 
   std::unique_ptr<TestableOutcome> outcome = constraint.createOutcome();
   EXPECT_TRUE(constraint.isSatisfied(state, outcome.get()));
@@ -130,9 +133,9 @@ TEST_F(CollisionFreeTest, AddPairwiseCheckPasses_IsSatisfied)
 
 TEST_F(CollisionFreeTest, AddPairwiseCheckFails_IsSatisfied)
 {
-  CollisionFree constraint(mStateSpace, mCollisionDetector);
+  CollisionFree constraint(mStateSpace, mSkeleton, mCollisionDetector);
 
-  auto state = mStateSpace->getScopedStateFromMetaSkeleton();
+  auto state = mStateSpace->getScopedStateFromMetaSkeleton(mSkeleton.get());
 
   constraint.addPairwiseCheck(mCollisionGroup1, mCollisionGroup3);
   CollisionFreeOutcome outcome;
@@ -154,9 +157,9 @@ TEST_F(CollisionFreeTest, AddPairwiseCheckFails_IsSatisfied)
 
 TEST_F(CollisionFreeTest, AddAndRemovePairwiseCheckFails_IsSatisfied)
 {
-  CollisionFree constraint(mStateSpace, mCollisionDetector);
+  CollisionFree constraint(mStateSpace, mSkeleton, mCollisionDetector);
 
-  auto state = mStateSpace->getScopedStateFromMetaSkeleton();
+  auto state = mStateSpace->getScopedStateFromMetaSkeleton(mSkeleton.get());
   // Also reuse this outcome object and ensure the correct result
   // is still returned,
   CollisionFreeOutcome outcome;
@@ -180,9 +183,9 @@ TEST_F(CollisionFreeTest, AddAndRemovePairwiseCheckFails_IsSatisfied)
 
 TEST_F(CollisionFreeTest, AddSelfCheckPasses_IsSatisfied)
 {
-  CollisionFree constraint(mStateSpace, mCollisionDetector);
+  CollisionFree constraint(mStateSpace, mSkeleton, mCollisionDetector);
 
-  auto state = mStateSpace->getScopedStateFromMetaSkeleton();
+  auto state = mStateSpace->getScopedStateFromMetaSkeleton(mSkeleton.get());
   constraint.addSelfCheck(mCollisionGroup1);
 
   std::unique_ptr<TestableOutcome> outcome = constraint.createOutcome();
@@ -192,9 +195,9 @@ TEST_F(CollisionFreeTest, AddSelfCheckPasses_IsSatisfied)
 
 TEST_F(CollisionFreeTest, AddSelfCheckFails_IsSatisfied)
 {
-  CollisionFree constraint(mStateSpace, mCollisionDetector);
+  CollisionFree constraint(mStateSpace, mSkeleton, mCollisionDetector);
 
-  auto state = mStateSpace->getScopedStateFromMetaSkeleton();
+  auto state = mStateSpace->getScopedStateFromMetaSkeleton(mSkeleton.get());
 
   constraint.addSelfCheck(mCollisionGroup3);
   CollisionFreeOutcome outcome;
@@ -214,9 +217,9 @@ TEST_F(CollisionFreeTest, AddSelfCheckFails_IsSatisfied)
 
 TEST_F(CollisionFreeTest, AddAndRemoveSelfCheckFails_IsSatisfied)
 {
-  CollisionFree constraint(mStateSpace, mCollisionDetector);
+  CollisionFree constraint(mStateSpace, mSkeleton, mCollisionDetector);
 
-  auto state = mStateSpace->getScopedStateFromMetaSkeleton();
+  auto state = mStateSpace->getScopedStateFromMetaSkeleton(mSkeleton.get());
 
   constraint.addSelfCheck(mCollisionGroup3);
   EXPECT_FALSE(constraint.isSatisfied(state));