+++ CHANGELOG.md
@@ -1,8 +1,27 @@
 ## Aikido 0 (prerelease)
 
-### 0.2.0 (201X-XX-XX)
+### 0.3.0 (201X-XX-XX)
+
+* Common
+  
+  * Fixed bug in StepSequence::getMaxSteps(): [#305](https://github.com/personalrobotics/aikido/pull/305)
 
 * State Space
+
+  * Refactored JointStateSpace and MetaSkeletonStateSpace: [#278](https://github.com/personalrobotics/aikido/pull/278)
+
+* Control
+
+  * Fixed CollisionGroup bugs in Hand executors: [#299](https://github.com/personalrobotics/aikido/pull/299)
+
+* Planner
+
+  * Added parabolic timing for linear spline [#302](https://github.com/personalrobotics/aikido/pull/302)
+
+### 0.2.0 (2018-01-09)
+
+* State Space
+
   * Moved MetaSkeletonStateSpaceSaver implementation to src: [#273](https://github.com/personalrobotics/aikido/pull/273)
 
 * Constraint
@@ -17,16 +36,17 @@
 
 * Planner
 
-  * Added World class: [#243](https://github.com/personalrobotics/aikido/pull/243), [#252](https://github.com/personalrobotics/aikido/pull/252)
+  * Added World class: [#243](https://github.com/personalrobotics/aikido/pull/243), [#252](https://github.com/personalrobotics/aikido/pull/252), [#265](https://github.com/personalrobotics/aikido/pull/265)
   * Added vector field planner [#246](https://github.com/personalrobotics/aikido/pull/246), [#262](https://github.com/personalrobotics/aikido/pull/262), [#268](https://github.com/personalrobotics/aikido/pull/268)
 
 * RViz
 
   * Added WorldInteractiveMarkerViewer: [#242](https://github.com/personalrobotics/aikido/pull/242)
+  * Added TrajectoryMarker: [#288](https://github.com/personalrobotics/aikido/pull/288)
 
 * IO
 
-  * Added support for new ErrorStr API in tinyxml2 6.0.0: [#290](https://github.com/personalrobotics/aikido/pull/290)
+  * Added support for new ErrorStr API in tinyxml2 6.0.0: [#290](https://github.com/personalrobotics/aikido/pull/290), [#295](https://github.com/personalrobotics/aikido/pull/295)
 
 * Build & Testing & ETC